Use seed_services_client for getting identity
SEED-909.

There are many similar cases in this file and across this project.
This is the smallest possible change I can make to be sure I'm doing
the right thing.

Use the API client to send requests to avoid duplicating code to
talk to the seed-identity-store API.

This uses new functionality in the seed_services_client to handle
retries.