Cache container image after installing dependencies
Before this commit, any change to the application would result in
`pip install` running (because the COPY had changed).

We only want to run the install again if the dependencies have changed,
so copy just the setup.py file to begin with and then cache the build
after installing packages. Then copy the rest of the app over.

This makes repeatedly building much faster for development.